More --install changes, improve non-SBo and sqf builds

+++ functions.d/installfunctions.sh
@@ -43,8 +43,8 @@ function install_packages
     elif ls /var/log/packages/"$pkgid"-* 1>/dev/null 2>/dev/null; then
       for instpkg in /var/log/packages/"$pkgid"-*; do
         if [ "$(basename "$instpkg" | rev | cut -f4- -d- | rev)" = "$pkgid" ]; then
-          log_verbose -a "A previous instance of $pkgid is already installed; upgrading ..."
-          if [ "$OPT_VERBOSE" = 'y' ]; then
+          log_normal -a "A previous instance of $pkgid is already installed; upgrading ..."
+          if [ "$OPT_VERYVERBOSE" = 'y' ]; then
             /sbin/upgradepkg --reinstall "$pkgpath" 2>&1 | tee -a "$ITEMLOG"
             stat=$?
           else
@@ -57,7 +57,7 @@ function install_packages
         fi
       done
     else
-      if [ "$OPT_VERBOSE" = 'y' ]; then
+      if [ "$OPT_VERBOSE" = 'y' -o "$OPT_INSTALL" = 'y' ]; then
         /sbin/installpkg --terse "$pkgpath" 2>&1 | tee -a "$MAINLOG" "$ITEMLOG"
         stat=$?
       else
@@ -84,7 +84,6 @@ function uninstall_packages
   local pkgpath pkgbase pkgid
   local etcnewfiles etcdirs etcfile etcdir
 
-  [ "$OPT_INSTALL" = 'y' ] && return 0
   [ "${HINT_no_uninstall[$itemid]}" = 'y' ] && return 0
 
   # Look for the package(s).
@@ -96,42 +95,46 @@ function uninstall_packages
   # Finally, look in the proper package repo
   [ "${#pkglist[@]}" = 0 ] && \
     pkglist=( $(ls "$SR_PKGREPO"/"$itemdir"/*.t?z 2>/dev/null) )
-  # should have something by now!
-  [ "${#pkglist[@]}" = 0 ] && \
-    { log_error -a "${itemid}: Can't find any packages to uninstall"; return 1; }
+  # oh well, never mind -- return quietly
+  [ "${#pkglist[@]}" = 0 ] && return 0
 
   for pkgpath in "${pkglist[@]}"; do
     pkgbase=$(basename "$pkgpath" | sed 's/\.t.z$//')
     pkgid=$(echo "$pkgbase" | rev | cut -f4- -d- | rev )
+
     # Is it installed?
     if [ -f /var/log/packages/"$pkgbase" ]; then
 
-      # Save a list of potential detritus in /etc
-      etcnewfiles=$(grep '^etc/.*\.new$' /var/log/packages/"$pkgbase")
-      etcdirs=$(grep '^etc/.*/$' /var/log/packages/"$pkgbase")
-
-      log_verbose -a "Uninstalling $pkgbase ..."
-      /sbin/removepkg "$pkgbase" >> "$ITEMLOG" 2>&1
-
-      # Remove any surviving detritus
-      for etcfile in $etcnewfiles; do
-        # (this is why we shouldn't run on an end user system!)
-        rm -f /"$etcfile" /"$(echo "$etcfile" | sed 's/\.new$//')"
-      done
-      for etcdir in $etcdirs; do
-        if [ -d "$etcdir" ]; then
-          find "$etcdir" -type d -depth -exec rmdir --ignore-fail-on-non-empty {} \;
+      if [ "$OPT_INSTALL" = 'y' ]; then
+        # Conventional gentle removepkg :-)
+        log_normal -a "Uninstalling $pkgbase ..."
+        /sbin/removepkg "$pkgbase" >> "$ITEMLOG" 2>&1
+      else
+        # Violent removal :D
+        # Save a list of potential detritus in /etc
+        etcnewfiles=$(grep '^etc/.*\.new$' /var/log/packages/"$pkgbase")
+        etcdirs=$(grep '^etc/.*/$' /var/log/packages/"$pkgbase")
+        # Run removepkg
+        log_verbose -a "Uninstalling $pkgbase ..."
+        /sbin/removepkg "$pkgbase" >> "$ITEMLOG" 2>&1
+        # Remove any surviving detritus
+        for etcfile in $etcnewfiles; do
+          rm -f /"$etcfile" /"$(echo "$etcfile" | sed 's/\.new$//')"
+        done
+        for etcdir in $etcdirs; do
+          if [ -d "$etcdir" ]; then
+            find "$etcdir" -type d -depth -exec rmdir --ignore-fail-on-non-empty {} \;
+          fi
+        done
+        # Do this last so it can mend things the package broke.
+        # The cleanup file can contain any required shell commands, for example:
+        #   * Reinstalling Slackware packages that conflict with the item's packages
+        #   * Unsetting environment variables set in an /etc/profile.d script
+        #   * Removing specific files and directories that removepkg doesn't remove
+        #   * Running depmod to remove references to removed kernel modules
+        if [ -n "${HINT_cleanup[$itemid]}" ]; then
+          . "${HINT_cleanup[$itemid]}" >> "$ITEMLOG" 2>&1
         fi
-      done
-
-      # Do this last so it can mend things the package broke.
-      # The cleanup file can contain any required shell commands, for example:
-      #   * Reinstalling Slackware packages that conflict with the item's packages
-      #   * Unsetting environment variables set in an /etc/profile.d script
-      #   * Removing specific files and directories that removepkg doesn't remove
-      #   * Running depmod to remove references to removed kernel modules
-      if [ -n "${HINT_cleanup[$itemid]}" ]; then
-        . "${HINT_cleanup[$itemid]}" >> "$ITEMLOG" 2>&1
       fi
 
     fi
